+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/handlers/v2/DescriptorHandler.java
@@ -1,14 +1,21 @@
 package com.sap.cloud.lm.sl.mta.handlers.v2;
 
-import com.sap.cloud.lm.sl.mta.handlers.ModulesSorter;
+import com.sap.cloud.lm.sl.common.util.Pair;
 import com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionDescriptor;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionModule;
+import com.sap.cloud.lm.sl.mta.model.v2.ExtensionProvidedDependency;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionRequiredDependency;
+import com.sap.cloud.lm.sl.mta.model.v2.ExtensionResource;
 import com.sap.cloud.lm.sl.mta.model.v2.Module;
+import com.sap.cloud.lm.sl.mta.model.v2.Platform;
+import com.sap.cloud.lm.sl.mta.model.v2.ModuleType;
+import com.sap.cloud.lm.sl.mta.model.v2.ResourceType;
+import com.sap.cloud.lm.sl.mta.model.v2.ProvidedDependency;
 import com.sap.cloud.lm.sl.mta.model.v2.RequiredDependency;
+import com.sap.cloud.lm.sl.mta.model.v2.Resource;
 
-public class DescriptorHandler extends com.sap.cloud.lm.sl.mta.handlers.v1.DescriptorHandler {
+public class DescriptorHandler {
 
     public RequiredDependency findRequiredDependency(DeploymentDescriptor descriptor, String moduleName, String dependencyName) {
         for (Module module : descriptor.getModules2()) {
@@ -21,7 +28,7 @@ public class DescriptorHandler extends com.sap.cloud.lm.sl.mta.handlers.v1.Descr
     }
 
     public ExtensionRequiredDependency findRequiredDependency(ExtensionDescriptor descriptor, String moduleName, String dependencyName) {
-        for (ExtensionModule module : descriptor.getModules2()) {
+        for (ExtensionModule module : descriptor.getModules()) {
             if (module.getName()
                 .equals(moduleName)) {
                 return findRequiredDependency(module, dependencyName);
@@ -50,10 +57,122 @@ public class DescriptorHandler extends com.sap.cloud.lm.sl.mta.handlers.v1.Descr
         return null;
     }
 
-    @Override
-    protected ModulesSorter getModuleSorter(com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor descriptor,
+    protected com.sap.cloud.lm.sl.mta.handlers.ModulesSorter getModuleSorter(DeploymentDescriptor descriptor,
         String parallelDeploymentProperty, String dependencyTypeProperty, String hardDependencyType) {
-        return new com.sap.cloud.lm.sl.mta.handlers.v2.ModulesSorter(descriptor, this, dependencyTypeProperty, hardDependencyType);
+        return new ModulesSorter(descriptor, this, dependencyTypeProperty, hardDependencyType);
+    }
+    
+    public ResourceType findPlatformResourceType(Platform platform, String resourceTypeName) {
+        for (ResourceType resourceType : platform.getResourceTypes2()) {
+            if (resourceType.getName()
+                .equals(resourceTypeName)) {
+                return resourceType;
+            }
+        }
+        return null;
     }
 
+    public ModuleType findPlatformModuleType(Platform platform, String moduleTypeName) {
+        for (ModuleType moduleType : platform.getModuleTypes2()) {
+            if (moduleType.getName()
+                .equals(moduleTypeName)) {
+                return moduleType;
+            }
+        }
+        return null;
+    }
+    
+    public Pair<Resource, ProvidedDependency> findDependency(DeploymentDescriptor descriptor, String dependencyName) {
+        Resource resource = findResource(descriptor, dependencyName);
+        if (resource != null) {
+            return new Pair<>(resource, null);
+        }
+        ProvidedDependency providedDependency = findProvidedDependency(descriptor, dependencyName);
+        if (providedDependency != null) {
+            return new Pair<>(null, providedDependency);
+        }
+        return null;
+    }
+    
+    public Resource findResource(DeploymentDescriptor descriptor, String resourceName) {
+        for (Resource resource : descriptor.getResources2()) {
+            if (resource.getName()
+                .equals(resourceName)) {
+                return resource;
+            }
+        }
+        return null;
+    }
+    
+    public ExtensionResource findResource(ExtensionDescriptor descriptor, String resourceName) {
+        for (ExtensionResource resource : descriptor.getResources()) {
+            if (resource.getName()
+                .equals(resourceName)) {
+                return resource;
+            }
+        }
+        return null;
+    }
+
+    public ProvidedDependency findProvidedDependency(DeploymentDescriptor descriptor, String providedDependencyName) {
+        for (Module module : descriptor.getModules2()) {
+            ProvidedDependency dependency = findProvidedDependency(module, providedDependencyName);
+            if (dependency != null) {
+                return dependency;
+            }
+        }
+        return null;
+    }
+    
+    public ExtensionProvidedDependency findProvidedDependency(ExtensionDescriptor descriptor, String providedDependencyName) {
+        for (ExtensionModule module : descriptor.getModules()) {
+            ExtensionProvidedDependency dependency = findProvidedDependency(module, providedDependencyName);
+            if (dependency != null) {
+                return dependency;
+            }
+        }
+        return null;
+    }
+    
+    public ProvidedDependency findProvidedDependency(Module module, String providedDependencyName) {
+        for (ProvidedDependency providedDependency : module.getProvidedDependencies2()) {
+            if (providedDependency.getName()
+                .equals(providedDependencyName)) {
+                return providedDependency;
+            }
+        }
+        return null;
+    }
+    
+    public ExtensionProvidedDependency findProvidedDependency(ExtensionModule module, String providedDependencyName) {
+        for (ExtensionProvidedDependency providedDependency : module.getProvidedDependencies2()) {
+            if (providedDependency.getName()
+                .equals(providedDependencyName)) {
+                return providedDependency;
+            }
+        }
+        return null;
+    }
+
+    public ExtensionModule findModule(ExtensionDescriptor descriptor, String moduleName) {
+        for (ExtensionModule module : descriptor.getModules()) {
+            if (module.getName()
+                .equals(moduleName)) {
+                return module;
+            }
+        }
+        return null;
+    }
+    
+    public Module findModule(DeploymentDescriptor descriptor, String moduleName) {
+        for (Module module : descriptor.getModules2()) {
+            if (module.getName()
+                .equals(moduleName)) {
+                return module;
+            }
+        }
+        return null;
+    }
+    
+    
 }
